More atan2 param casts.
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/trunk@2911 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/vector3d.h
@@ -275,7 +275,7 @@ namespace core
 		{
 			vector3d<T> angle;
 
-			const f64 tmp = (atan2(X, Z) * RADTODEG64);
+			const f64 tmp = (atan2((f64)X, (f64)Z) * RADTODEG64);
 			angle.Y = (T)tmp;
 
 			if (angle.Y < 0)
@@ -309,7 +309,7 @@ namespace core
 			{
 				if (X!=0)
 				{
-					angle.Y = (T)(atan2(Z,X) * RADTODEG64);
+					angle.Y = (T)(atan2((f64)Z,(f64)X) * RADTODEG64);
 				}
 				else if (Z<0)
 					angle.Y=180;
